requirements.txt
annotated-types==0.7.0
astroid==3.2.4
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==1.0.1
bs4==0.0.2
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.41.0
cloudpathlib==0.20.0
confection==0.1.5
coverage==7.6.1
cymem==2.0.10
dill==0.3.8
dnspython==2.7.0
Flask==3.0.3
flask-paginate==2024.4.12
gunicorn==23.0.0
h11==0.14.0
idna==3.8
inflect==7.4.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
MongoSanitizer==0.0.1
more-itertools==10.5.0
murmurhash==1.0.11
numpy==2.0.2
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
platformdirs==4.3.2
preshed==3.0.9
psycopg2==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pylint==3.2.7
pymongo==4.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rich==13.9.4
selenium==4.24.0
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.3.2
tomlkit==0.13.2
Top==1.0.0
tqdm==4.67.1
trio==0.26.2
trio-websocket==0.11.1
typeguard==4.4.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
Werkzeug==3.0.4
wrapt==1.17.0
wsproto==1.2.0
